+++ R/weightit2cbps.R
@@ -60,7 +60,7 @@
 #'     \item{`maxit`}{the maximum number of iterations for convergence of the optimization. Passed to the `control` argument of `optim()`. Default is 1000.
 #'     }
 #'     \item{`quantile`}{
-#'     A named list of quantiles (values between 0 and 1) for each continuous covariate, which are used to create additional variables that when balanced ensure balance on the corresponding quantile of the variable. For example, setting `quantile = list(x1 = c(.25, .5. , .75))` ensures the 25th, 50th, and 75th percentiles of `x1` in each treatment group will be balanced in the weighted sample. Can also be a single number (e.g., `.5`) or an unnamed list of length 1 (e.g., `list(c(.25, .5, .75))`) to request the same quantile(s) for all continuous covariates, or a named vector (e.g., `c(x1 = .5, x2 = .75`) to request one quantile for each covariate. Only allowed with binary and multi-category treatments.
+#'     A named list of quantiles (values between 0 and 1) for each continuous covariate, which are used to create additional variables that when balanced ensure balance on the corresponding quantile of the variable. For example, setting `quantile = list(x1 = c(.25, .5. , .75))` ensures the 25th, 50th, and 75th percentiles of `x1` in each treatment group will be balanced in the weighted sample. Can also be a single number (e.g., `.5`) or an unnamed list of length 1 (e.g., `list(c(.25, .5, .75))`) to request the same quantile(s) for all continuous covariates, or a named vector (e.g., `c(x1 = .5, x2 = .75)` to request one quantile for each covariate. Only allowed with binary and multi-category treatments.
 #'   }
 #' }
 #'
@@ -165,10 +165,9 @@ weightit2cbps <- function(covs, treat, s.weights, estimand, focal, subset,
     covs <- add_missing_indicators(covs)
   }
 
-  covs <- cbind(covs, .int_poly_f(covs, poly = moments, int = int, center = TRUE))
-
-  covs <- cbind(covs, .quantile_f(covs, qu = A[["quantile"]], s.weights = s.weights,
-                                  focal = focal, treat = treat))
+  covs <- cbind(.int_poly_f(covs, poly = moments, int = int, center = TRUE),
+                .quantile_f(covs, qu = A[["quantile"]], s.weights = s.weights,
+                            focal = focal, treat = treat))
 
   t.lev <- get_treated_level(treat)
   treat <- binarize(treat, one = t.lev)
@@ -177,7 +176,7 @@ weightit2cbps <- function(covs, treat, s.weights, estimand, focal, subset,
   covs <- covs[, colnames(covs) %nin% colinear.covs.to.remove, drop = FALSE]
 
   mod_covs <- svd(cbind(`(Intercept)` = 1, scale(covs)))$u
-  bal_covs <- svd(cbind(`(Intercept)` = 1, scale(covs)))$u
+  bal_covs <- mod_covs
 
   over <- if_null_then(A$over, FALSE)
   chk::chk_flag(over)
@@ -354,16 +353,15 @@ weightit2cbps.multi <- function(covs, treat, s.weights, estimand, focal, subset,
     covs <- add_missing_indicators(covs)
   }
 
-  covs <- cbind(covs, .int_poly_f(covs, poly = moments, int = int, center = TRUE))
-
-  covs <- cbind(covs, .quantile_f(covs, qu = A[["quantile"]], s.weights = s.weights,
-                                  focal = focal, treat = treat))
+  covs <- cbind(.int_poly_f(covs, poly = moments, int = int, center = TRUE),
+                .quantile_f(covs, qu = A[["quantile"]], s.weights = s.weights,
+                            focal = focal, treat = treat))
 
   colinear.covs.to.remove <- setdiff(colnames(covs), colnames(make_full_rank(covs)))
   covs <- covs[, colnames(covs) %nin% colinear.covs.to.remove, drop = FALSE]
 
   mod_covs <- svd(cbind(`(Intercept)` = 1, scale(covs)))$u
-  bal_covs <- svd(cbind(`(Intercept)` = 1, scale(covs)))$u
+  bal_covs <- mod_covs
 
   over <- if_null_then(A$over, FALSE)
   chk::chk_flag(over)
@@ -595,16 +593,16 @@ weightit2cbps.cont <- function(covs, treat, s.weights, subset, missing, moments,
     covs <- add_missing_indicators(covs)
   }
 
-  covs <- cbind(covs, .int_poly_f(covs, poly = moments, int = int, center = TRUE, orthogonal_poly = TRUE))
+  covs <- .int_poly_f(covs, poly = moments, int = int, center = TRUE, orthogonal_poly = TRUE)
 
   colinear.covs.to.remove <- setdiff(colnames(covs), colnames(make_full_rank(covs)))
   covs <- covs[, colnames(covs) %nin% colinear.covs.to.remove, drop = FALSE]
 
-  treat <- (treat - w.m(treat, s.weights)) / sqrt(col.w.v(treat, s.weights))
+  treat <- scale_w(treat, s.weights)
 
-  covs <- scale(covs,
-                center = col.w.m(covs, s.weights),
-                scale = sqrt(col.w.v(covs, s.weights)))
+  for (i in seq_col(covs)) {
+    covs[,i] <- scale_w(covs[,i], s.weights)
+  }
 
   mod_covs <- cbind(`(Intercept)` = 1,
                     scale(svd(covs)$u))
@@ -782,14 +780,15 @@ weightitMSM2cbps <- function(covs.list, treat.list, s.weights, subset, missing, 
       covs.list[[i]] <- add_missing_indicators(covs.list[[i]])
     }
 
-    covs.list[[i]] <- cbind(covs.list[[i]], .int_poly_f(covs.list[[i]], poly = moments,
-                                                        int = int, center = TRUE))
-
-    if (treat.types[i] %in% c("binary", "multi") &&
-        is_not_null(A[["quantile"]])) {
-      covs.list[[i]] <- cbind(covs.list[[i]], .quantile_f(covs.list[[i]], qu = A[["quantile"]],
-                                                          s.weights = s.weights,
-                                                          treat = treat.list[[i]]))
+    if (treat.types[i] %in% c("binary", "multi")) {
+      covs.list[[i]] <- cbind(.int_poly_f(covs.list[[i]], poly = moments,
+                                          int = int, center = TRUE),
+                              .quantile_f(covs.list[[i]], qu = A[["quantile"]],
+                                          s.weights = s.weights))
+    }
+    else {
+      covs.list[[i]] <- .int_poly_f(covs.list[[i]], poly = moments,
+                                    int = int, center = TRUE)
     }
 
     colinear.covs.to.remove <- setdiff(colnames(covs.list[[i]]), colnames(make_full_rank(covs.list[[i]])))
@@ -799,13 +798,15 @@ weightitMSM2cbps <- function(covs.list, treat.list, s.weights, subset, missing, 
       treat.types[i],
       "binary" = binarize(treat.list[[i]], one = get_treated_level(treat.list[[i]])),
       "multi" = factor(treat.list[[i]]),
-      "continuous"= (treat.list[[i]] - w.m(treat.list[[i]], s.weights)) / sqrt(col.w.v(treat.list[[i]], s.weights))
+      "continuous" = scale_w(treat.list[[i]], s.weights)
     )
 
+    for (j in seq_col(covs.list[[i]])) {
+      covs.list[[i]][,j] <- scale_w(covs.list[[i]][,j], s.weights)
+    }
+
     covs.list[[i]] <- cbind(`(Intercept)` = 1,
-                            scale(svd(scale(covs.list[[i]],
-                                            center = col.w.m(covs.list[[i]], s.weights),
-                                            scale = sqrt(col.w.v(covs.list[[i]], s.weights))))$u))
+                            scale(svd(covs.list[[i]])$u))
   }
 
   over <- if_null_then(A$over, FALSE)